BioVossEncoder.BinarySequenceMatrix

BioVossEncoder.binary_sequence_matrix

BioVossEncoder.binaryseq

BioVossEncoder.BinarySequenceMatrix — Type

struct BinarySequenceMatrix{A<:NucleicAcidAlphabet, B<:BitMatrix}

The BinarySequenceMatrix struct represents a binary matrix encoding a sequence of nucleic acids. This is also called the Voss representation of a sequence.

Fields

alphabet::A: The nucleic acid alphabet used for encoding.

bsm::B: The binary matrix representing the encoded sequence.

Constructors

BinarySequenceMatrix(sequence::SeqOrView{A}): Constructs a BinarySequenceMatrix from a sequence of nucleic acids.

Arguments

sequence::SeqOrView{A}: The sequence of nucleic acids to be encoded.

BioVossEncoder.binary_sequence_matrix — Method

binary_sequence_matrix(sequence::SeqOrView{A}) where {A <: NucleicAcidAlphabet}

Create a binary sequence matrix from a given nucleic acid sequence.

Arguments

sequence: A nucleic acid sequence.

Returns

The binary sequence matrix.

BioVossEncoder.binaryseq — Method

binaryseq(sequence::SeqOrView{A}, molecule::Union{DNA,RNA}) where {A <: Alphabet}
binaryseq(sequence::SeqOrView{A}, molecule::Tuple{T}) where {A <: Alphabet, T <: BioSymbol}

Converts a sequence of nucleotides into a binary representation.

Arguments

sequence::SeqOrView{A}: The input sequence of nucleotides.

`molecule::BioSymbol: The nucleotide to be encoded as 1, while others are encoded as 0.

molecules::Tuple{Vararg{T}}: The nucleotides to be encoded as 1, while others are encoded as 0.

Returns

A BitVector representing the binary encoding of the input sequence, where 1 indicates the presence of the specified nucleotide and 0 indicates the absence.